requirements.txt
langchain
python-dotenv
langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain_ollama
langchain_groq
langchain_openai
pandas
numpy
arxiv
wikipedia
ddgs
